+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/handlers/LoadMoreHandler.kt
@@ -16,38 +16,51 @@
 
 package io.getstream.chat.android.compose.handlers
 
-import androidx.compose.foundation.lazy.LazyColumn
+import androidx.annotation.VisibleForTesting
 import androidx.compose.foundation.lazy.LazyListState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.derivedStateOf
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
+import androidx.compose.runtime.snapshotFlow
+import kotlinx.coroutines.flow.distinctUntilChanged
 
 /**
- * Handler to be used with [LazyColumn] to implement infinite scroll.
+ * Handler to notify more items should be loaded when the user scrolls to the end of the list.
  *
  * @param listState The state of the list used to control scrolling.
- * @param loadMoreThreshold The number if items before the end of the list.
+ * @param loadMoreThreshold The number if items before the end of the list. Default is half of the visible items.
  * @param loadMore Handler for load more action.
  */
 @Composable
 public fun LoadMoreHandler(
     listState: LazyListState,
-    loadMoreThreshold: Int = 3,
+    loadMoreThreshold: () -> Int = { listState.layoutInfo.visibleItemsInfo.size / 2 },
     loadMore: () -> Unit,
 ) {
-    val shouldLoadMore by remember {
-        derivedStateOf {
+    LaunchedEffect(listState) {
+        snapshotFlow {
             val totalItemsCount = listState.layoutInfo.totalItemsCount
-            val lastVisibleItem = listState.layoutInfo.visibleItemsInfo.lastOrNull()
-                ?: return@derivedStateOf false
-
-            lastVisibleItem.index > (totalItemsCount - loadMoreThreshold - 1)
+            val lastVisibleItemIndex = listState.layoutInfo.visibleItemsInfo.lastOrNull()?.index ?: -1
+            shouldLoadMore(
+                totalItemsCount = totalItemsCount,
+                lastVisibleItemIndex = lastVisibleItemIndex,
+                loadMoreThreshold = loadMoreThreshold(),
+            )
         }
-    }
-
-    LaunchedEffect(shouldLoadMore) {
-        if (shouldLoadMore) { loadMore() }
+            .distinctUntilChanged()
+            .collect { shouldLoadMore ->
+                if (shouldLoadMore) {
+                    loadMore()
+                }
+            }
     }
 }
+
+@VisibleForTesting
+internal fun shouldLoadMore(
+    totalItemsCount: Int,
+    lastVisibleItemIndex: Int,
+    loadMoreThreshold: Int,
+): Boolean =
+    totalItemsCount > 0 && // List isn’t empty
+        totalItemsCount > loadMoreThreshold && // Ensure list is large enough
+        lastVisibleItemIndex > totalItemsCount - loadMoreThreshold - 1